+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - Introduce `intellijPlatform.caching` extension
 - Allow for customizing the cached IntelliJ Platform directory name with `intellijPlatform.caching.ides.name`
 - Introduce `intellijPlatform.caching.ides.enabled` to control IntelliJ Platform caching globally
+- Integrate Reporting API to `VerifyPluginProjectConfigurationTask` by @alajemba-vik JetBrains/intellij-platform-gradle-plugin#2016
+- Integrate Problems API to improve error reporting in `VerifyPluginTask` by @alajemba-vik JetBrains/intellij-platform-gradle-plugin#1987
 
 ### Changed
 
@@ -20,6 +22,7 @@
 - Correctly resolve `~` in `GradleProperties` to the users home directory.
 - Remove `extendsFrom` usages in custom IntelliJ Platform test configurations to avoid propagation of dependencies on bundled modules and plugins coming from the base IntelliJ Platform.
 - Refactor `IntelliJPlatformDependenciesExtension` to replace lambda-based configuration with `Action` for improved interoperability and Groovy support  JetBrains/intellij-platform-gradle-plugin#2017.
+- Refactor `LocalIvyArtifactPathComponentMetadataRule` to validate module type and version against `productInfo` when resolving Ivy artifacts. JetBrains/intellij-platform-gradle-plugin#2015
 
 ## [2.7.2] - 2025-08-14